Moving all changes to utils.py
